requirements.txt
aio_pika==9.5.5
aiormq==6.8.1
anyio==4.8.0
anyio==4.9.0
colorama==0.4.6
fastapi==0.115.14
httpx==0.28.1
pydantic==2.11.7
python-dotenv==1.1.1
python_jose==3.4.0
redis==5.2.1
SQLAlchemy==2.0.38
starlette==0.47.1
typing_extensions==4.14.0
uuid7==0.1.0
xxhash==3.5.0
yarl==1.18.3
yarl==1.19.0
